secretmanager/modules/src/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-logging
google-cloud-secret-manager
google-api-python-client
oauth2client